chore: move help types to types.ts
