--- frontend/src/components/help/help-types.ts
@@ -1,15 +1,0 @@
-export type QAItem = {
-  id: string;
-  question: string;
-  answer: React.ReactNode;
-  tags: string[];
-};
-
-export enum EnumQATags {
-  SMS = 'sms',
-  MAIL = 'mail',
-  REK_MAIL = 'rek-mail',
-  RECIPIENTS = 'recipients',
-  DOCUMENTS = 'documents',
-  SENDER = 'sender',
-}
